fix: fix a bug where converting Int to Int16 caused some devices to crash (why not all?)
feat: add independent psbt signer view controller and ability for app to auto open when a .psbt is airdropped and then sign/process/broadcast that psbt or export it after updating it

fix: force bitcoind to not rescan when creating a single-sig fully noded wallet

feat: add new info button for segueing directly to active wallet details

+++ BitSense/Wallet Logic/CreateFullyNodedWallet.swift
@@ -55,7 +55,7 @@ class Keys {
         return fingerprint
     }
     
-    class func bip84AccountXpub(masterKey: String, coinType: String, account: Int) -> String? {
+    class func bip84AccountXpub(masterKey: String, coinType: String, account: Int16) -> String? {
         var xpub:String?
         if let mk = HDKey(masterKey) {
             if let path = BIP32Path("m/84'/\(coinType)'/\(account)'") {
